update changelog, docstring, move to utils
